+++ python/blickfeld_scanner/stream/point_cloud.py
@@ -12,6 +12,7 @@ from __future__ import print_function
 import math
 import gzip
 import time
+import warnings
 from google.protobuf.internal.decoder import _DecodeVarint32
 from google.protobuf.internal.encoder import _EncodeVarint
 
@@ -29,7 +30,6 @@ terminate = False
 def signal_handler(sig, frame):
     global terminate
     terminate = True
-
 
 # Catch CTRL-C and terminate and close the point cloud stream
 signal.signal(signal.SIGINT, signal_handler)
@@ -90,20 +90,34 @@ class point_cloud(object):
 
     :param connection: connection to the device
     :type connection: :py:class:`blickfeld_scanner.scanner.connection`
+    :param from_file: path to file, of which to stream a point cloud, if this is set, no connection should be given
+    :param filter: DEPRECATED
+
+        > Introduced in BSL v2.10 and firmware v1.9
+
+        Filter points and returns by point attributes during the post-processing on the device. Is replaced by 'point_filter'.
+    :param reference_frame:
+        > Introduced in BSL v2.10 and firmware v1.9
+
+        Frame representing the desired data. To request a field, set it to any value (also in submessages). For a repeated field, add at least one element.
+    :param point_filter:
+        > Introduced in BSL v2.10 and firmware v1.9
+
+        Filter points and returns by point attributes during the post-processing on the device. This replaces the 'filter' parameter
+        
+    :param extend_subscribe_request:
+        > Introduced in BSL v2.13 and firmware v1.13
+
+        Extend point cloud subscription request with additional parameters. This is mainly used internally.
     """
 
-    """ Reference frame: XYZ coordinates """
-    REF_FRAME_XYZ = REF_FRAME_XYZ
-    """ Reference frame: XYZ coordinates, intensity """
-    REF_FRAME_XYZ_I = REF_FRAME_XYZ_I
-    """ Reference frame: XYZ coordinates, intensity, frame id, scanline id, point id, return id """
-    REF_FRAME_XYZ_I_ID = REF_FRAME_XYZ_I_ID
-    """ Reference frame: XYZ coordinates, intensity, frame id, scanline id, point id, return id, timestamps """
-    REF_FRAME_XYZ_I_ID_TS = REF_FRAME_XYZ_I_ID_TS
-    """ Reference frame: ambient_light_level, intensity, range, frame id, scanline id, point id """
-    REF_FRAME_DEPTH_MAP = REF_FRAME_DEPTH_MAP
+    REF_FRAME_XYZ = REF_FRAME_XYZ  #: Reference frame: XYZ coordinates
+    REF_FRAME_XYZ_I = REF_FRAME_XYZ_I  #: Reference frame: XYZ coordinates, intensity
+    REF_FRAME_XYZ_I_ID = REF_FRAME_XYZ_I_ID  #: Reference frame: XYZ coordinates, intensity, frame id, scanline id, point id, return id
+    REF_FRAME_XYZ_I_ID_TS = REF_FRAME_XYZ_I_ID_TS  #: Reference frame: XYZ coordinates, intensity, frame id, scanline id, point id, return id, timestamps
+    REF_FRAME_DEPTH_MAP = REF_FRAME_DEPTH_MAP  #: Reference frame: ambient_light_level, intensity, range, frame id, scanline id, point id
 
-    def __init__(self, connection=None, from_file=None, filter=None, reference_frame=None):
+    def __init__(self, connection=None, from_file=None, filter=None, reference_frame=None, point_filter=None, extend_subscribe_request=None):
         self._metadata = point_cloud_pb2.PointCloud.Metadata()
 
         if connection and from_file:
@@ -125,10 +139,17 @@ class point_cloud(object):
 
             req = connection_pb2.Request()
             req.subscribe.point_cloud.SetInParent()
+            if filter and point_filter:
+                raise Exception("Either provide 'filter' or 'point_filter'.\nPlease use 'point_filter' instead of 'filter'. 'filter' is deprecated.")
+            if point_filter:
+                req.subscribe.point_cloud.filter.CopyFrom(point_filter)
             if filter:
+                warnings.warn("Please use 'point_filter' instead of 'filter'. 'filter' is deprecated.", DeprecationWarning, stacklevel=2)
                 req.subscribe.point_cloud.filter.CopyFrom(filter)
             if reference_frame:
                 req.subscribe.point_cloud.reference_frame.CopyFrom(reference_frame)
+            if extend_subscribe_request:
+                req.subscribe.point_cloud.MergeFrom(extend_subscribe_request)
 
             self._metadata.header.device.CopyFrom(self._connection.send_request(req).event.point_cloud.header)
 
@@ -139,8 +160,11 @@ class point_cloud(object):
             # Open file
             self._file_name = from_file
             self._file = gzip.open(self._file_name, 'rb')
-            self._file.seek(0, 2)
-            self._file_size = self._file.tell()
+            try:
+                self._file.seek(0, 2)
+                self._file_size = self._file.tell()
+            except EOFError as e:
+                self._file_size = None
             self._file.seek(0, 0)
 
             # Read header
@@ -149,16 +173,17 @@ class point_cloud(object):
             self._metadata.header.ParseFromString(buf_header)
             self._stream_buffered = False
 
-            # Find footer offset
-            while self._file_size > self._file.tell():
-                offset_footer = self._file.tell()
-                self._get_file_block()
-
-            # Read footer
-            self._file.seek(offset_footer)
-            if self._read_stream():
-                self._metadata.footer.CopyFrom(self._stream_data.footer)
-                self._stream_buffered = False
+            if self._file_size:
+                # Find footer offset
+                while self._file_size > self._file.tell():
+                    offset_footer = self._file.tell()
+                    self._get_file_block()
+    
+                # Read footer
+                self._file.seek(offset_footer)
+                if self._read_stream():
+                    self._metadata.footer.CopyFrom(self._stream_data.footer)
+                    self._stream_buffered = False
             self._file.seek(offset_data)
 
     def __del__(self):
@@ -189,13 +214,16 @@ class point_cloud(object):
         return buf
 
     def _read_stream(self):
-        if self._file_size == self._file.tell():
+        if self._file_size and self._file_size == self._file.tell():
             self._stream_buffered = False
             return False
 
         self._stream_data.Clear()
-        self._stream_data.ParseFromString(self._get_file_block())
-        self._stream_buffered = True
+        try:
+            self._stream_data.ParseFromString(self._get_file_block())
+            self._stream_buffered = True
+        except EOFError as e:
+            self._stream_buffered = False
         return self._stream_buffered
 
     def get_metadata(self):
@@ -281,7 +309,12 @@ class point_cloud(object):
 
         :param file_name: Path to the file where it should be dumped
         :param compresslevel: The compresslevel argument is an integer from 0 to 9 controlling the level of compression; 1 is fastest and produces the least compression, and 9 is slowest and produces the most compression. 0 is no compression. The default is 1. If frames are lost during the recording decrease the compression level.
+
+        .. deprecated:: 2.13.0
+            Since BSL v2.13.0 this function is deprecated, because of performance issues. Please use :py:func:`blickfeld_scanner.scanner.scanner.record_point_cloud_stream` for a performance improved recording.
         """
+        warnings.warn("This function is deprecated, because of performance issues. Please use scanner.record_point_cloud_stream for a performance improved recording.", DeprecationWarning, stacklevel=2)
+
         if self._ofile:
             raise Exception("The output file has already been opened. To open another file, please call 'stop_recording' first to close the current output file.")
         self._prev_scan_pattern_str = b""
@@ -300,3 +333,15 @@ class point_cloud(object):
 
             self._ofile.close()
             self._ofile = None
+
+    def stop(self):
+        """ Stop and unsubscribe of the stream
+        """
+        self.stop_recording()
+        
+        if self._connection:
+            req = connection_pb2.Request()
+            req.unsubscribe.point_cloud.SetInParent()
+            self._connection.send(req)
+        else:
+            raise NotImplementedError("File stream not supported.")